App statefulness overhaul
- Remove App `stateful` property
- Add Item `state_key` property

Items will always be read and updated, but a `dismissed` item will only be un-dismissed
if the state key has changed. This allows the script to update the
underlying data without resurfacing the item if the changes aren't worth
bringing it back into the foreground.

Apps that do not emit state keys on items will continue to use the
previous logic that resurfaces an item where any of the daat fields have
changed.